checkpoint refactor of issuance
