+++ packages/verite/lib/issuer/credential-fulfillment.ts
@@ -1,4 +1,3 @@
-import { isString } from "lodash"
 import { v4 as uuidv4 } from "uuid"
 
 import {
@@ -7,7 +6,6 @@ import {
   Issuer,
   CredentialPayload,
   DidKey,
-  JWT,
   Attestation,
   CredentialManifest,
   ClaimFormat
@@ -15,75 +13,19 @@ import {
 import {
   CREDENTIAL_RESPONSE_TYPE_NAME,
   VC_CONTEXT_URI,
-  VERIFIABLE_CREDENTIAL_TYPE_NAME,
   VERIFIABLE_PRESENTATION_TYPE_NAME,
   VERITE_VOCAB_URI
 } from "../utils"
 import {
-  encodeVerifiableCredential,
   encodeVerifiablePresentation
 } from "../utils/credentials"
 
 import type {
-  CreateCredentialOptions,
   CreatePresentationOptions
 } from "did-jwt-vc/src/types"
+import { buildVerifiableCredential, signVerifiableCredential } from "./credential"
 
 const DEFAULT_CONTEXT = [VC_CONTEXT_URI, { "@vocab": VERITE_VOCAB_URI }]
-/**
- * Build a VerifiableCredential containing an attestation for the given holder.
- */
-export async function buildAndSignVerifiableCredential(
-  signer: Issuer,
-  subject: string | DidKey,
-  attestation: Attestation | Attestation[],
-  credentialType: string | string[],
-  payload: Partial<CredentialPayload> = {},
-  options?: CreateCredentialOptions
-): Promise<JWT> {
-  const subjectId = parseSubjectId(subject)
-
-  // append all types other than "Verifiable Credential", which is already there
-  const finalCredentialTypes = [VERIFIABLE_CREDENTIAL_TYPE_NAME].concat(
-    Array.isArray(credentialType)
-      ? credentialType.filter((c) => c !== VERIFIABLE_CREDENTIAL_TYPE_NAME)
-      : credentialType !== VERIFIABLE_CREDENTIAL_TYPE_NAME
-      ? [credentialType]
-      : []
-  )
-
-  // For attestations, preserve the array or object structure
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let attsns: any[] | any
-  if (Array.isArray(attestation)) {
-    attsns = attestation.map((att) => {
-      return {
-        id: subjectId,
-        [att["type"].toString()]: att
-      }
-    })
-  } else {
-    attsns = {
-      id: subjectId,
-      [attestation["type"].toString()]: attestation
-    }
-  }
-  const vcPayload = Object.assign(
-    {
-      "@context": DEFAULT_CONTEXT,
-      type: finalCredentialTypes,
-      credentialSubject: attsns,
-      issuanceDate: new Date(),
-      issuer: { id: signer.did }
-    },
-    payload
-  )
-  return encodeVerifiableCredential(vcPayload, signer, options)
-}
-
-function parseSubjectId(subject: string | DidKey) {
-  return isString(subject) ? subject : subject.subject
-}
 
 /**
  * Build a VerifiablePresentation containing a list of attestations.
@@ -99,20 +41,19 @@ export async function buildAndSignFulfillment(
   payload: Partial<CredentialPayload> = {},
   options?: CreatePresentationOptions
 ): Promise<EncodedCredentialFulfillment> {
-  const subjectId = parseSubjectId(subject)
-  const encodedCredentials = await buildAndSignVerifiableCredential(
+  const encodedCredentials = await buildVerifiableCredential(
     signer,
-    subjectId,
+    subject,
     attestation,
     credentialType,
-    payload,
-    options
+    payload
   )
+  const signedCredential = await signVerifiableCredential(signer, encodedCredentials, options)
   const format = ClaimFormat.JwtVc
 
   const encodedPresentation = await encodeVerifiablePresentation(
     signer.did,
-    encodedCredentials,
+    signedCredential,
     signer,
     options,
     [VERIFIABLE_PRESENTATION_TYPE_NAME, CREDENTIAL_RESPONSE_TYPE_NAME],
@@ -168,3 +109,7 @@ export async function buildAndSignMultiVcFulfillment(
 
   return encodedPresentation as unknown as EncodedCredentialFulfillment
 }
+function parseSubjectId(subject: string | DidKey) {
+  throw new Error("Function not implemented.")
+}
+
