+++ packages/verite/lib/issuer/index.ts
@@ -1,3 +1,4 @@
+export * from "./credential"
 export * from "./credential-application"
 export * from "./credential-offer"
 export * from "./credential-fulfillment"
